kafka_consumer/requirements.txt
kafka-python==2.0.2
requests==2.31.0
pytz==2022.7
psycopg2==2.9.3